+++ SmartPower/Core/Configurator.cs
@@ -3,7 +3,6 @@
 // License: Public Domain
 
 using Bindito.Core;
-using IgorZ.TimberDev.CustomInstantiator;
 using TimberApi.ConfiguratorSystem;
 using TimberApi.SceneSystem;
 using Timberborn.Attractions;
@@ -12,13 +11,14 @@ using Timberborn.MechanicalSystem;
 using Timberborn.PowerGenerating;
 using IgorZ.TimberDev.Utils;
 using Timberborn.Workshops;
+using UnityDev.Utils.LogUtilsLite;
 
 // ReSharper disable InconsistentNaming
 // ReSharper disable UnusedMember.Local
 // ReSharper disable once CheckNamespace
 namespace IgorZ.SmartPower {
 
-[Configurator(SceneEntrypoint.InGame)]
+[Configurator(SceneEntrypoint.All)]
 // ReSharper disable once UnusedType.Global
 sealed class Configurator : IConfigurator {
   static readonly PrefabPatcher.RequiredComponentsDep SmartAttractionDeps =
@@ -28,6 +28,9 @@ sealed class Configurator : IConfigurator {
   static readonly string PatchId = typeof(Configurator).FullName;
 
   public void Configure(IContainerDefinition containerDefinition) {
+    if (Features.DebugExVerboseLogging && DebugEx.LoggingSettings.VerbosityLevel < 5) {
+      DebugEx.LoggingSettings.VerbosityLevel = 5;
+    }
     CustomizableInstantiator.AddPatcher(
         PatchId,
         prefab => {