+++ SmartPower/Core/SmartPoweredAttraction.cs
@@ -2,9 +2,10 @@
 // Author: igor.zavoychinskiy@gmail.com
 // License: Public Domain
 
-using Timberborn.BaseComponentSystem;
 using Timberborn.EnterableSystem;
 using Timberborn.MechanicalSystem;
+using Timberborn.Persistence;
+using Timberborn.TickSystem;
 using UnityDev.Utils.LogUtilsLite;
 
 // ReSharper disable once CheckNamespace
@@ -12,33 +13,50 @@ namespace IgorZ.SmartPower {
 
 /// <summary>Smart version of the stock game powered attraction.</summary>
 /// <remarks>It only consumes power if there are attendees in the building.</remarks>
-public sealed class SmartPoweredAttraction : BaseComponent {
-  Enterable _enterable;
+public sealed class SmartPoweredAttraction : TickableComponent, IPostInitializableLoadedEntity {
+
+  /// <summary>Indicates if the power consumption has been disabled due to no attendees.</summary>
+  // ReSharper disable once MemberCanBePrivate.Global
+  public bool ConsumptionDisabled { get; private set; }
+
+  #region IPostInitializableLoadedEntity implementation
+
+  /// <inheritdoc/>
+  public void PostInitializeLoadedEntity() {
+    UpdateConsumingState();
+  }
+
+  #endregion
+
+  #region TickableComponent overrides
+
+  /// <inheritdoc/>
+  public override void Tick() {
+    UpdateConsumingState();
+  }
+
+  #endregion
+
+  #region Implementation
+
   MechanicalBuilding _mechanicalBuilding;
-  
+  Enterable _enterable;
+
   void Awake() {
     _mechanicalBuilding = GetComponentFast<MechanicalBuilding>();
     _enterable = GetComponentFast<Enterable>();
-    _enterable.EntererAdded += (_, _) => UpdateConsumingState();
-    _enterable.EntererRemoved += (_, _) => UpdateConsumingState();
-  }
-
-  void Start() {
-    UpdateConsumingState();
   }
 
   void UpdateConsumingState() {
     var newState = _enterable.NumberOfEnterersInside == 0;
-    if (newState == _mechanicalBuilding.ConsumptionDisabled) {
-      return;
+    if (newState != ConsumptionDisabled) {
+      HostedDebugLog.Fine(this, newState ? "Disable power consumption" : "Enable power consumption");
+      ConsumptionDisabled = newState;
     }
-    if (newState) {
-      HostedDebugLog.Fine(this, "Disable power consumption");
-    } else {
-      HostedDebugLog.Fine(this, "Enable power consumption");
-    }
-    _mechanicalBuilding.ConsumptionDisabled = newState;
+    _mechanicalBuilding.ConsumptionDisabled = ConsumptionDisabled;
   }
+
+  #endregion
 }
 
 }