+++ TimberCommons/IrrigationSystemUI/GoodConsumingBuildingFragmentPatch.cs
@@ -1,0 +1,29 @@
+﻿// Timberborn Mod: Timberborn Commons
+// Author: igor.zavoychinskiy@gmail.com
+// License: Public Domain
+
+using HarmonyLib;
+using IgorZ.TimberDev.UI;
+using Timberborn.GoodConsumingBuildingSystem;
+using Timberborn.GoodConsumingBuildingSystemUI;
+using Timberborn.Localization;
+using UnityEngine.UIElements;
+
+// ReSharper disable InconsistentNaming
+namespace IgorZ.TimberCommons.IrrigationSystemUI {
+
+/// Harmony patch to show supply in days and hours.
+[HarmonyPatch(typeof(GoodConsumingBuildingFragment), nameof(GoodConsumingBuildingFragment.UpdateProgressBar))]
+static class GoodConsumingBuildingFragmentPatch {
+
+  // ReSharper disable once UnusedMember.Local
+  static void Postfix(bool __runOriginal, ILoc ____loc, Label ____hoursLeft,
+                      GoodConsumingBuilding ____goodConsumingBuilding) {
+    if (!__runOriginal) {
+      return;  // The other patches must follow the same style to properly support the skip logic!
+    }
+    ____hoursLeft.text = CommonFormats.DaysHoursFormat(____loc, ____goodConsumingBuilding.HoursUntilNoSupply);
+  }
+}
+
+}