+++ TimberCommons/IrrigationSystemUI/GrowableToolPanelItemFactoryPatch.cs
@@ -1,0 +1,26 @@
+﻿// Timberborn Mod: Timberborn Commons
+// Author: igor.zavoychinskiy@gmail.com
+// License: Public Domain
+
+using HarmonyLib;
+using IgorZ.TimberDev.UI;
+using Timberborn.Growing;
+using Timberborn.GrowingUI;
+using Timberborn.Localization;
+using UnityEngine.UIElements;
+
+// ReSharper disable InconsistentNaming
+namespace IgorZ.TimberCommons.IrrigationSystemUI {
+
+[HarmonyPatch(typeof(GrowableToolPanelItemFactory), nameof(GrowableToolPanelItemFactory.Create))]
+static class GrowableToolPanelItemFactoryPatch {
+  // ReSharper disable once UnusedMember.Local
+  static void Postfix(Growable growable, bool __runOriginal, ref VisualElement __result, ILoc ____loc) {
+    if (!__runOriginal) {
+      return;  // The other patches must follow the same style to properly support the skip logic!
+    }
+    __result.Q<Label>("GrowthTime").text = CommonFormats.DaysHoursFormat(____loc, growable.GrowthTimeInDays * 24f);
+  }
+}
+
+}