+++ TimberCommons/WaterValveComponent/WaterValve.cs
@@ -2,7 +2,6 @@
 // Author: igor.zavoychinskiy@gmail.com
 // License: Public Domain
 
-using System;
 using Bindito.Core;
 using IgorZ.TimberCommons.WaterService;
 using Timberborn.BlockSystem;
@@ -26,43 +25,54 @@ public sealed class WaterValve : TickableComponent, IPersistentEntity, IFinished
   // ReSharper disable InconsistentNaming
 
   [SerializeField]
+  [Tooltip("Indicates if the valve UI fragment should be shown in game for this building.")]
   bool _showUIPanel = true;
 
   [SerializeField]
+  [Tooltip("Relative coordinates at which the water will be taken.")]
   Vector2Int _inputCoordinates = new(0, 0);
 
   [SerializeField]
+  [Tooltip("Relative coordinates at which the water will be dropped.")]
   Vector2Int _outputCoordinates = new(0, 2);
 
   [SerializeField]
+  [Tooltip("Maximum water flow in cubic metres per second.")]
   float _waterFlowPerSecond = 1.5f;
 
   [SerializeField]
+  [Tooltip("Indicates if the flow rate can be adjust in game by the player.")]
   bool _canChangeFlowInGame = true;
 
   [SerializeField]
+  [Tooltip("If the flow change is allowed, then this will be the minimum possible value to set.")]
   float _minimumInGameFlow = 0;
 
   [SerializeField]
+  [Tooltip("Don't take water at input below this threshold. Set to -1 to disable this check.")]
   float _minimumWaterLevelAtIntake = 0.2f;
 
   [SerializeField]
+  [Tooltip("Don't drop water at output above this threshold. Set to -1 to disable this check.")]
   float _maximumWaterLevelAtOuttake = 0.6f;
 
   [SerializeField]
+  [Tooltip("Indicates that water level check at the output can be disabled via UI.")]
   bool _allowDisablingOutputLevelCheck = true;
 
   [SerializeField]
+  [Tooltip("Tells if the contamination should also be moved.")]
   bool _moveContaminatedWater = true;
 
   [SerializeField]
+  [Tooltip("If set, then will start on non-zero flow and stop if no water is being moving via the valve.")]
   ParticleSystem _particleSystem = null;
 
   // ReSharper restore InconsistentNaming
   #endregion
 
   #region API
-  /// <summary>Absolute height of the water surface above the map at teh valve intake.</summary>
+  /// <summary>Absolute height of the water surface above the map at the valve intake.</summary>
   public float WaterHeightAtInput { get; private set; }
 
   /// <summary>Absolute height of the water surface above the map at the valve outtake.</summary>
@@ -204,9 +214,6 @@ public sealed class WaterValve : TickableComponent, IPersistentEntity, IFinished
   #region TickableComponent implemenatation
   /// <inheritdoc/>
   public override void StartTickable() {
-    if (!_directWaterServiceAccessor.IsValid) {
-      throw new InvalidOperationException("WaterValve requires operational DirectWaterServiceAccessor. See the logs!");
-    }
     base.StartTickable();
 
     _valveBaseZ = _blockObject.Coordinates.z;