+++ SmartPower/package/CHANGELOG.md
@@ -1,3 +1,14 @@
+# v1.7 (1 Feb 2024)
+* [Fix #27] Smart Power does not work for buildings with supply.
+* [FIx #30] Manufactories without working places are handled incorrectly.
+
+# v1.6 (8 Dec 2023)
+* [Fix #20] Fix crashes on game load in Timberborn 0.5.4.2.
+
+# v1.5 (5 Dec 2023)
+* Experimental version for Update 5.
+* [Fix #18] Fix crashes on Timberborn 0.5.4.0.
+
 # v1.4 (27 Aug 2023)
 * [Fix #8] Optimize UI fragments handling.